move files into dlc2action_annotation
