+++ dlc2action_annotation/colors.txt
