+++ dlc2action_annotation/icons/hand.png
