+++ dlc2action_annotation/icons/pantone.png
