+++ dlc2action_annotation/icons/trash.png
