+++ dlc2action_annotation/widgets/canvas.py
@@ -6,7 +6,7 @@
 from PyQt5.QtCore import QTimer
 from vispy.scene import SceneCanvas
 
-from utils import SignalEmitter, WorkerThread, read_stack
+from dlc2action_annotation.utils import SignalEmitter, WorkerThread, read_stack
 
 from .viewbox import VideoViewBox, VideoViewBox3D
