+++ dlc2action_annotation/widgets/settings.py
