+++ dlc2action_annotation/widgets/viewbox.py
@@ -10,7 +10,7 @@ import numpy as np
 import vispy
 from vispy.scene import SceneCanvas
 
-from utils import SignalEmitter
+from dlc2action_annotation.utils import SignalEmitter
 
 from .visuals import AnimalMarkers, BoxVisual, Markers3d
